# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Mision
# during dev, should point to your local machine
host: localhost:8080
# basePath prefixes all resource paths 
basePath: /ms
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  #Log in with facebook
  /login/facebook:
    # binds a127 app logic to a route
    x-swagger-router-controller: login
    post:
      description: Lgo in with facebook
      # used as the method name of the controller
      operationId: loginWithFacebook
      parameters:
        - name: facebookToken
          in: body
          description: user's facebook token
          required: true
          schema:
            properties:
              facebookToken:
                type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/LoginResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  #All Users
  /users:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    get:
      description: Returns all users to the caller (testing purpose only)
      # used as the method name of the controller
      operationId: getAllUsers
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UsersResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /user/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    get:
      description: Returns one user of the given id to the caller
      # used as the method name of the controller
      operationId: getUserByID
      parameters:
        - name: auth
          in: header
          description: current user's auth token
          required: true
          type: string
        - name: id
          in: path
          description: The id of the user
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UserResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /user/me:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    get:
      description: Returns all user profiles to the caller
      # used as the method name of the controller
      operationId: getMyProfile
      parameters:
        - name: auth
          in: header
          description: current user's auth token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UserResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /missions:
    x-swagger-router-controller: mission
    get:
      description: Returns all missions to the caller
      # used as the method name of the controller
      operationId: allMissions
      parameters:
        - name: auth
          in: header
          description: current user's auth token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MissionsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /mission/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: mission
    get:
      description: Returns one mission of the given id to the caller
      # used as the method name of the controller
      operationId: getMissionByID
      parameters:
        - name: auth
          in: header
          description: current user's auth token
          required: true
          type: string
        - name: id
          in: path
          description: The id of the mission
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MissionResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /mission/create:
    # binds a127 app logic to a route
    x-swagger-router-controller: mission
    post:
      description: Create new mission
      # used as the method name of the controller
      operationId: createMission
      parameters:
        - name: auth
          in: header
          description: current user's auth token
          required: true
          type: string
        - name: mission
          in: body
          description: mission info
          required: true
          schema:
            properties:
              title:
                type: string
              description:
                type: string
              price:
                type: number
              dueDate:
                type: string
              lat:
                type: string
              lng:
                type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MissionResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /mission/{id}/update:
    # binds a127 app logic to a route
    x-swagger-router-controller: mission
    post:
      description: Update existing mission
      # used as the method name of the controller
      operationId: editMission
      parameters:
        - name: auth
          in: header
          description: current user's auth token
          required: true
          type: string
        - name: id
          in: path
          description: The id of the mission
          required: true
          type: number
        - name: mission
          in: body
          description: mission info
          required: true
          schema:
            properties:
              title:
                type: string
              description:
                type: string
              price:
                type: number
              dueDate:
                type: string
              lat:
                type: string
              lng:
                type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MissionResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  LoginResponse:
    required:
      - token
    properties:
      token:
        type: string
  UsersResponse:
    required:
      - users
    properties:
      users:
        type: array
        items: 
          type: object
          properties: 
            id: 
              type: number
            facebookId:
              type: string
            firstname:
              type: string
            lastname:
              type: string
            gender:
              type: string
            birthday:
              type: string
            email:
              type: string
            avatar:
              type: string
            hasCar:
              type: boolean
            job:
              type: string
            country:
              type: string
            state:
              type: string
            city:
              type: string
            street:
              type: string
            postalCode:
              type: string
            lat:
              type: number
            lng:
              type: number
  UserResponse:
    required:
      - user
    properties:
      user:
        type: object
        properties:
          id: 
            type: number
          facebookId:
            type: string
          firstname:
            type: string
          lastname:
            type: string
          gender:
            type: string
          birthday:
            type: string
          email:
            type: string
          avatar:
            type: string
          hasCar:
            type: boolean
          job:
            type: string
          country:
            type: string
          state:
            type: string
          city:
            type: string
          street:
            type: string
          postalCode:
            type: string
          lat:
            type: number
          lng:
            type: number
  MissionsResponse:
    required:
      - missions
    properties:
      missions:
        type: array
        items:
          type: object
          properties:
            id: 
              type: number
            title:
              type: string
            description:
              type: string
            status:
              type: number
            price:
              type: number
            authorID:
              type: number
            accepterID:
              type: number
            postDate:
              type: string
            dueDate:
              type: string
            lat:
              type: number
            lng:
              type: number
            country:
              type: string
            state:
              type: string
            city:
              type: string
            street:
              type: string
            postalCode:
              type: string
  MissionResponse:
    required:
      - mission
    properties:
      mission:
        type: object
        properties:
          id: 
            type: number
          title:
            type: string
          description:
            type: string
          status:
            type: number
          price:
            type: number
          authorID:
            type: number
          accepterID:
            type: number
          postDate:
            type: string
          dueDate:
            type: string
          lat:
            type: number
          lng:
            type: number
          country:
            type: string
          state:
            type: string
          city:
            type: string
          street:
            type: string
          postalCode:
            type: string
  GeneralResponse:
    required:
      - message
    properties:
      message:
        type: string